Move function to external file
